restructure base blocks [#2]
