+++ critter/blocks/misc.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+
+
+class BranchRateModel(BaseModel):
+    """ <branchRateModel/>"""
+
+    def __str__(self):
+        return self.xml
+
+    @property
+    def xml(self):
+        return "<branchRateModel></branchRateModel>"
